Important!
메인 페이지에서 바로 홈페이지를 불러오는 것으로 바꾸고, 상품을 등록하면 새로운 홈페이지를 생성하지 않도록 pushreplacement를 pop으로 바꾸어 수정했습니다. 더불어, 가격 세자릿수마다 콤마가 표시되도록 pubspec.yaml에 intl: ^0.18.0를 추가하고 코드 일부를 수정하였습니다.

+++ lib/models/product.dart
@@ -1,10 +1,9 @@
-import 'package:flutter/foundation.dart';
 
 class Product {
   final String image;
   final String name;
   final String description;
-  final double price;
+  final int price;
   
   Product({
     required this.image,
